+++ xazz-compiler/src/policy/printer.rs
@@ -98,10 +98,9 @@ fn print_layer(layer: &LayerKind) -> String {
             out_channels,
             kernel_size,
         } => format!("Conv1d({}, {})", out_channels, kernel_size),
-        LayerKind::Embedding {
-            vocab_size,
-            embed_dim,
-        } => format!("Embedding({}, {})", vocab_size, embed_dim),
+        LayerKind::Embedding { vocab, embed_dim } => {
+            format!("Embedding({}, {})", vocab.display(), embed_dim)
+        }
     }
 }